animation responsive when i rezise my screen
