+++ src/components/pages/home/Home.jsx
@@ -13,7 +13,8 @@ import "./Home.css";
 import Button from "../../button/Button";
 import { DataContext } from "../../../context/DataProvider";
 import OverlayRevealContent from "../../animation/OverlayRevealContent";
-import Experience from "../../animation/Experience";
+import TitleAnimation from "../../animation/TitleAnimation";
+import ProfilAnimation from "../../animation/ProfilAnimation";
 
 export default function Home() {
   const { owner } = useContext(DataContext);
@@ -33,10 +34,10 @@ export default function Home() {
       className="relative m-auto h-full flex flex-col gap-14 items-center overflow-hidden lg:flex-row"
     >
       <div className="fixed h-[200%] w-[100%] -rotate-[15deg] bg-mainColorContrast -top-[50%] -left-[83%] hidden lg:block"></div>
-      <Experience animatedElement={"profil"} />
+      <ProfilAnimation />
       <div className="flex flex-col items-center w-full lg:w-2/4 lg:ml-[40%]">
         <div className="relative flex flex-col items-center justify-center h-[600px] w-4/5 -top-5">
-          <Experience animatedElement={"title"} />
+          <TitleAnimation />
           <p className="my-5 text-md lg:text-xl lg:w-[80%]">
             Je suis concepteur web ainsi que développeur front-end basé à la
             frontière franco-suisse. Je suis passionné par la création
